Changed doubles to floats in FastExpSketch

+++ Data Analyser/FastExpSketch.cpp
@@ -21,13 +21,13 @@ void FastExpSketch::initialize()
         permInit[i] = i + 1;
     }
 
-    M = std::vector<double>(size);
+    M = std::vector<float>(size);
     for(size_t i = 0; i < size; i++)
     {
-        M[i] = std::numeric_limits<double>::infinity();
+        M[i] = std::numeric_limits<float>::infinity();
     }
 
-    maxValue = std::numeric_limits<double>::infinity();
+    maxValue = std::numeric_limits<float>::infinity();
 }
 
 template <typename T>
@@ -38,29 +38,29 @@ void swapValues(T& lhs, T& rhs)
     rhs = temp;
 }
 
-// Function calculating hash value of i || k, returning double value in range [0, 1]
-double FastExpSketch::hash(unsigned i, unsigned k, unsigned seed = 17)
+// Function calculating hash value of i || k, returning float value in range [0, 1]
+float FastExpSketch::hash(unsigned i, unsigned k, unsigned seed = 17)
 {
     std::bitset<8 * sizeof(i)> bitsetI(i);
     std::bitset<8 * sizeof(k)> bitsetK(k);
     std::bitset<8 * sizeof(seed)> bitsetSeed(seed);
     std::string concatHash = bitsetI.to_string() + bitsetK.to_string() + bitsetSeed.to_string();
     std::hash<std::string> hash{};
-    return static_cast<double>((hash(concatHash)) / pow(2, 8 * sizeof(size_t)));
+    return static_cast<float>((hash(concatHash)) / pow(2, 8 * sizeof(size_t)));
 }
 
 // Function updating data sketch on arrival of pair (i, lambda)
-void FastExpSketch::update(unsigned i, double lambda)
+void FastExpSketch::update(unsigned i, float lambda)
 {
     std::mt19937 mt{i};
-    double S = 0;
+    float S = 0;
     bool updateMax = false;
     std::vector<unsigned> P = permInit;
 
     for(unsigned k = 1; k <= size; k++)
     {
-        double U = hash(i, k);
-        double E = -log(U) / lambda;
+        float U = hash(i, k);
+        float E = -log(U) / lambda;
         S += E / (size - k + 1);
         if (S > maxValue)
         {
@@ -72,15 +72,15 @@ void FastExpSketch::update(unsigned i, double lambda)
         swapValues(P[k-1], P[r-1]);
         size_t j = P[k-1] - 1;
 
-        if (M[j] == std::numeric_limits<double>::infinity() && maxValue == std::numeric_limits<double>::infinity()) 
+        if (M[j] == std::numeric_limits<float>::infinity() && maxValue == std::numeric_limits<float>::infinity()) 
         {
             updateMax = true;
         }
-        else if (M[j] == std::numeric_limits<double>::infinity() || maxValue == std::numeric_limits<double>::infinity()) 
+        else if (M[j] == std::numeric_limits<float>::infinity() || maxValue == std::numeric_limits<float>::infinity()) 
         {
             updateMax = false;
         }
-        else if (fabs(M[j] - maxValue) < 10 * std::numeric_limits<double>::epsilon()) 
+        else if (fabs(M[j] - maxValue) < 10 * std::numeric_limits<float>::epsilon()) 
         {
             updateMax = true;
         }
@@ -95,9 +95,9 @@ void FastExpSketch::update(unsigned i, double lambda)
 }
 
 // Function calculating estimation of the sketch cardinality
-double FastExpSketch::estimateCardinality()
+float FastExpSketch::estimateCardinality()
 {
-    double sum = 0.0;
+    float sum = 0.0;
     for(const auto& m : M)
     {
         sum += m;
@@ -107,5 +107,5 @@ double FastExpSketch::estimateCardinality()
     {
         return 0.0;
     }
-    return (double)(size - 1) / sum;
+    return (float)(size - 1) / sum;
 }
